Significant improvement to AddRangeAsync for deserializing expando objects.
Added new custom serialization process to centralize serialization with wrappers and to enforce the not mapped attribute to be respected during serialization.

Created multiple caching dictionaries. This allows after the first discovery of an attribute, mapped location, or more, it will be cached and not require rediscovery on follow up use.

Removed all use case of Newtonsoft JSON serialization.